refactor(gui): useCellUrlConfig composable
