+++ packages/nc-gui/composables/index.ts
@@ -24,3 +24,4 @@ export * from './useSmartsheetStore'
 export * from './useLTARStore'
 export * from './useExpandedFormStore'
 export * from './useSharedFormViewStore'
+export * from './useCellUrlConfig'
